Modularize validators
